pyproject.toml
[tool.poetry]
name = "irreg"
version = "0.1.0"
description = ""
authors = ["Stanislau Trukhan <stas.truhan@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1"
h5py = "^3.1.0"
Pillow = "^8.1.0"
matplotlib = "^3.3.4"
imagesize = "^1.2.0"
numpy = "^1.20.1"
SimpleITK = "^2.0.2"
pandas = "^1.2.2"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipython = "^7.20.0"
notebook = "^6.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
